\name{MOTcorrelationAssigner-package}
\alias{MOTcorrelationAssigner-package}
\alias{MOTcorrelationAssigner}
\docType{package}
\title{
Performs the single sample classification using the PAM centroids and the gene expression values
}
\description{
This package is a part of MOTassigner pipeline. It classifiers the samples using the pam centroids score
}
\details{
\tabular{ll}{
Package: \tab MOTcorrelationAssigner\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-03-16\cr
License: \tab GPL-2\cr
}
	single sample correlation , sample info file
}
\author{
Pawan Poudel
Maintainer: Pawan Poudel, <paawanpoudel@gmail.com>
}
\references{
MOTassinger, 2016
}
\keyword{ package }

